Refactor stock schema and remove websocket/Massive data flow

+++ tradepulseai-backend/stock-service/src/main/java/com/tradepulseai/stockservice/repository/ExchangeRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface ExchangeRepository extends JpaRepository<Exchange, Long> {
+public interface ExchangeRepository extends JpaRepository<Exchange, Integer> {
     Optional<Exchange> findByMic(String mic);
 }
